Make polygon

Makes a polygon from a point set

poly_make(point_set, step_size = 1)

Arguments

point_set

a polygon in array or matrix form, in which the first two columns contain x and y vertex coordinates.

step_size

a numeric indicating the spacing of the points

Value

A data frame containing coordinates for polygon vertices and the slice it originated from (if present in the original point set).