Subset by structures

Subset a segmentation class object to only contain certain structures

seg_sub_str(segmentation, structures, planes_chosen = NULL)

Arguments

segmentation

an object of class segmentation

structures

a vector of character structure IDs

planes_chosen

a vector of planes to be further subset. Defaults to NULL, meaning the 3 anatomical planes will be considered.

Value

A segmentation class object containing only the structures defined in structures and, optionally, the planes defined in planes_chosen. The metadata are recompiled based on the new composition of the object.